refactor unit tests specific to view methods
